refactor(logic): reorganise prolog related files into dedicated pkg
