+++ x/logic/predicate/string_test.go
@@ -205,68 +205,89 @@ func TestStringBytes(t *testing.T) {
 			query       string
 			wantError   error
 			wantSuccess bool
-		}{
-			// inspired from https://github.com/SWI-Prolog/swipl-devel/blob/V9.1.21/src/Tests/core/test_string.pl#L91
+		}{ /*
+				// inspired from https://github.com/SWI-Prolog/swipl-devel/blob/V9.1.21/src/Tests/core/test_string.pl#L91
+				{
+					query:       "string_bytes(aap, [97, 97, 112], ascii).",
+					wantSuccess: true,
+				},
+				{
+					program:     `test :- string_bytes(aap, B, utf8), B == [97, 97, 112].`,
+					query:       "test.",
+					wantSuccess: true,
+				},
+				{
+					program:     `test :- string_bytes(S, [97, 97, 112], utf8), S == "aap".`,
+					query:       "test.",
+					wantSuccess: true,
+				},
+				{
+					program:     `test :- string_bytes(aap, B, 'utf-16be'), B == [0, 97, 0, 97, 0, 112].`,
+					query:       "test.",
+					wantSuccess: true,
+				},
+				{
+					program:     `test :- string_bytes(S, [0, 97, 0, 97, 0, 112], 'utf-16be'), S == "aap".`,
+					query:       "test.",
+					wantSuccess: true,
+				},
+				{
+					program:     `test :- string_bytes(aap, B, 'utf-16le'), B ==[97, 0, 97, 0, 112, 0].`,
+					query:       "test.",
+					wantSuccess: true,
+				},
+				{
+					program:     `test :- string_bytes(S, [97, 0, 97, 0, 112, 0], 'utf-16le'), S == "aap".`,
+					query:       "test.",
+					wantSuccess: true,
+				},
+				{
+					program:     `test :- string_bytes(今日は, B, utf8), B == [228,187,138,230,151,165,227,129,175].`,
+					query:       "test.",
+					wantSuccess: true,
+				},
+				{
+					program:     `test :- string_bytes(S, [228,187,138,230,151,165,227,129,175], utf8), S == "今日は".`,
+					query:       "test.",
+					wantSuccess: true,
+				},
+				{
+					program:     `test :- string_bytes(今日は, B, 'utf-16le'), B == [202,78,229,101,111,48].`,
+					query:       "test.",
+					wantSuccess: true,
+				},
+				{
+					program:     `test :- string_bytes(S, [202,78,229,101,111,48], 'utf-16le'), S == "今日は".`,
+					query:       "test.",
+					wantSuccess: true,
+				},
+				// error cases
+
+				{
+					query:       `string_bytes(_, [202,78,229,101,111,48], foo).`,
+					wantSuccess: false,
+					wantError:   fmt.Errorf("string_bytes/3: invalid encoding: foo"),
+				},*/
 			{
-				query:       "string_bytes(aap, [97, 97, 112], ascii).",
-				wantSuccess: true,
-			},
-			{
-				program:     `test :- string_bytes(aap, B, utf8), B == [97, 97, 112].`,
-				query:       "test.",
-				wantSuccess: true,
-			},
-			{
-				program:     `test :- string_bytes(S, [97, 97, 112], utf8), S == "aap".`,
-				query:       "test.",
-				wantSuccess: true,
-			},
-			{
-				program:     `test :- string_bytes(aap, B, 'utf-16be'), B == [0, 97, 0, 97, 0, 112].`,
-				query:       "test.",
-				wantSuccess: true,
-			},
-			{
-				program:     `test :- string_bytes(S, [0, 97, 0, 97, 0, 112], 'utf-16be'), S == "aap".`,
-				query:       "test.",
-				wantSuccess: true,
-			},
-			{
-				program:     `test :- string_bytes(aap, B, 'utf-16le'), B ==[97, 0, 97, 0, 112, 0].`,
-				query:       "test.",
-				wantSuccess: true,
-			},
-			{
-				program:     `test :- string_bytes(S, [97, 0, 97, 0, 112, 0], 'utf-16le'), S == "aap".`,
-				query:       "test.",
-				wantSuccess: true,
-			},
-			{
-				program:     `test :- string_bytes(今日は, B, utf8), B == [228,187,138,230,151,165,227,129,175].`,
-				query:       "test.",
-				wantSuccess: true,
-			},
-			{
-				program:     `test :- string_bytes(S, [228,187,138,230,151,165,227,129,175], utf8), S == "今日は".`,
-				query:       "test.",
-				wantSuccess: true,
-			},
-			{
-				program:     `test :- string_bytes(今日は, B, 'utf-16le'), B == [202,78,229,101,111,48].`,
-				query:       "test.",
-				wantSuccess: true,
-			},
-			{
-				program:     `test :- string_bytes(S, [202,78,229,101,111,48], 'utf-16le'), S == "今日は".`,
-				query:       "test.",
-				wantSuccess: true,
-			},
-			// error cases
-			{
-				query:       `string_bytes(_, [202,78,229,101,111,48], foo).`,
+				query:       `string_bytes(_, _, foo).`,
 				wantSuccess: false,
-				wantError:   fmt.Errorf("string_bytes/3: invalid encoding: foo"),
-			},
+				wantError:   fmt.Errorf("string_bytes/3: error(instantiation_error,string_bytes/3)"),
+			}, /*
+				{
+					query:       `string_bytes(_, wtf, utf8).`,
+					wantSuccess: false,
+					wantError:   fmt.Errorf("string_bytes/3: error(type_error(list,wtf),string_bytes/3)"),
+				},
+				{
+					query:       `string_bytes(foo(bar), _, utf8).`,
+					wantSuccess: false,
+					wantError:   fmt.Errorf("string_bytes/3: invalid compound term: expected a list of character_code or integer"),
+				},
+				{
+					query:       `string_bytes(_, foo(bar), utf8).`,
+					wantSuccess: false,
+					wantError:   fmt.Errorf("string_bytes/3: error(type_error(list,foo(bar)),string_bytes/3)"),
+				},*/
 		}
 		for nc, tc := range cases {
 			Convey(fmt.Sprintf("Given the query #%d: %s", nc, tc.query), func() {
@@ -285,6 +306,9 @@ func TestStringBytes(t *testing.T) {
 
 							Convey("When the predicate is called", func() {
 								sols, err := interpreter.QueryContext(ctx, tc.query)
+								Reset(func() {
+									So(sols.Close(), ShouldBeNil)
+								})
 
 								Convey("Then the error should be nil", func() {
 									So(err, ShouldBeNil)
